Reduzi o código adicionando uma view para cada mesa com computadores da proaluno
