+++ app/Http/Controllers/SociaisController.php
@@ -3,10 +3,11 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use app\Http\Models;
 
 class SociaisController extends Controller
 {
     public function index(){
-        return view('sociais');
+        return view('sociais.sociais');
     }
 }
