+++ resources/views/sociais/mesas/mesa2.blade.php
@@ -1,0 +1,65 @@
+//mesa 2
+var rectangle = svgContainer.append("rect")
+      .attr("x", 450)
+      .attr("y", 10)
+      .attr('fill', 'white')
+      .attr("width", 100)
+      .attr("height", 200);
+
+            //computadores da esquerda
+            var circle = svgContainer.append('circle')
+            .attr('cx', 465)
+            .attr('cy', 25)
+            .attr('r', 10)
+            .attr('stroke', 'black')
+            .attr('fill', 'red');
+
+            var circle = svgContainer.append('circle')
+            .attr('cx', 465)
+            .attr('cy', 75)
+            .attr('r', 10)
+            .attr('stroke', 'black')
+            .attr('fill', 'red');
+
+            var circle = svgContainer.append('circle')
+            .attr('cx', 465)
+            .attr('cy', 125)
+            .attr('r', 10)
+            .attr('stroke', 'black')
+            .attr('fill', 'red');
+
+            var circle = svgContainer.append('circle')
+            .attr('cx', 465)
+            .attr('cy', 175)
+            .attr('r', 10)
+            .attr('stroke', 'black')
+            .attr('fill', 'red');
+
+            //computadores da direita
+            var circle = svgContainer.append('circle')
+            .attr('cx', 525)
+            .attr('cy', 25)
+            .attr('r', 10)
+            .attr('stroke', 'black')
+            .attr('fill', 'red');
+
+            var circle = svgContainer.append('circle')
+            .attr('cx', 525)
+            .attr('cy', 75)
+            .attr('r', 10)
+            .attr('stroke', 'black')
+            .attr('fill', 'red');
+
+            var circle = svgContainer.append('circle')
+            .attr('cx', 525)
+            .attr('cy', 125)
+            .attr('r', 10)
+            .attr('stroke', 'black')
+            .attr('fill', 'red');
+
+            var circle = svgContainer.append('circle')
+            .attr('cx', 525)
+            .attr('cy', 175)
+            .attr('r', 10)
+            .attr('stroke', 'black')
+            .attr('fill', 'red');
