+++ resources/views/sociais/mesas/mesa4.blade.php
@@ -1,0 +1,65 @@
+//mesa 4
+var rectangle = svgContainer.append("rect")
+      .attr("x", 1050)
+      .attr("y", 10)
+      .attr('fill', 'white')
+      .attr("width", 100)
+      .attr("height", 200);
+
+            //computadores da esquerda
+            var circle = svgContainer.append('circle')
+            .attr('cx', 1065)
+            .attr('cy', 25)
+            .attr('r', 10)
+            .attr('stroke', 'black')
+            .attr('fill', 'red');
+
+            var circle = svgContainer.append('circle')
+            .attr('cx', 1065)
+            .attr('cy', 75)
+            .attr('r', 10)
+            .attr('stroke', 'black')
+            .attr('fill', 'red');
+
+            var circle = svgContainer.append('circle')
+            .attr('cx', 1065)
+            .attr('cy', 125)
+            .attr('r', 10)
+            .attr('stroke', 'black')
+            .attr('fill', 'red');
+
+            var circle = svgContainer.append('circle')
+            .attr('cx', 1065)
+            .attr('cy', 175)
+            .attr('r', 10)
+            .attr('stroke', 'black')
+            .attr('fill', 'red');
+
+            //computadores da direita
+            var circle = svgContainer.append('circle')
+            .attr('cx', 1125)
+            .attr('cy', 25)
+            .attr('r', 10)
+            .attr('stroke', 'black')
+            .attr('fill', 'red');
+
+            var circle = svgContainer.append('circle')
+            .attr('cx', 1125)
+            .attr('cy', 75)
+            .attr('r', 10)
+            .attr('stroke', 'black')
+            .attr('fill', 'red');
+
+            var circle = svgContainer.append('circle')
+            .attr('cx', 1125)
+            .attr('cy', 125)
+            .attr('r', 10)
+            .attr('stroke', 'black')
+            .attr('fill', 'red');
+
+            var circle = svgContainer.append('circle')
+            .attr('cx', 1125)
+            .attr('cy', 175)
+            .attr('r', 10)
+            .attr('stroke', 'black')
+            .attr('fill', 'red');
